refacto reducers
