+++ src/client/reducers/game/functions.js
@@ -1,0 +1,41 @@
+
+const reducerEmitGameStatus = (state, action, initialState) => {
+  switch (action.gameStatus) {
+    case 'Start': {
+      return {
+        ...state,
+        items: action.payload,
+        start: true,
+        gameIsStarted: true,
+        pause: false
+      }
+    }
+    case 'Pause': {
+      return {
+        ...state,
+        items: action.payload,
+        start: false,
+        pause: true,
+        gameIsStarted: true,
+      }
+    }
+    case 'Stop': {
+      return initialState
+    }
+    default: return state
+  }
+}
+
+const reducerEmitGamePieces = (state, action) => {
+  return state
+}
+
+const reducerGameFlow = (state, action) => {
+  return state
+}
+
+export {
+  reducerEmitGameStatus,
+  reducerEmitGamePieces,
+  reducerGameFlow,
+}
