+++ src/client/reducers/reducer.js
@@ -2,10 +2,10 @@ import { combineReducers } from 'redux'
 import { reducer as formReducer } from 'redux-form'
 
 import alert from './alert'
-import user from './user'
-import games from './games'
-import game from './game'
-import tetrimino from './tetrimino'
+import user from './user/user'
+import games from './games/games'
+import game from './game/game'
+import tetrimino from './tetrimino/tetrimino'
 
 const appReducer = combineReducers({
   user,
