+++ tests/client/components/Room/PlayGround.test.js
@@ -22,7 +22,7 @@ describe('>>>>PLAYGROUND - REACT-REDUX (Shallow + passing the {store} directly',
   let wrapper;
   beforeEach(() => {
     const store = mockStore(initialState)
-    wrapper = shallow(<ConnectedPlayground game={game} store={store}/>).dive()
+    wrapper = shallow(<ConnectedPlayground game={game} store={store}/>)
   })
   it('renders without crashing', () => {
     expect(wrapper.length).to.equal(1)
